Validate uri contains blob endpoint
Checking before attempting to get a reference as it would not come back when an internal directory was passed.
Also moved the delete from storage call to be after the database save.

+++ NetCoreAngular/Helpers/StorageHelper.cs
@@ -58,9 +58,12 @@ namespace NetCoreAngular.Helpers
         {
             CloudStorageAccount account = CloudStorageAccount.Parse(storageAccount);
 
-            CloudBlockBlob blockBlob = new CloudBlockBlob(new Uri(uri), account.Credentials);
+            if (uri.Contains(account.BlobEndpoint.ToString()))
+            {
+                CloudBlockBlob blockBlob = new CloudBlockBlob(new Uri(uri), account.Credentials);
 
-            await blockBlob.DeleteIfExistsAsync();
+                await blockBlob.DeleteIfExistsAsync();
+            }
         }
     }
 }
